update new york data
